Implement two endpoint as requirement

+++ PlayStudioQuestEngine/QuestEngine.Domain/Entities/Milestone.cs
@@ -4,8 +4,8 @@
     {
         public int Index { get; set; }
         public string Title { get; set; }
-        public int RequiredPoint { get; set; }
-        public int Reward { get; set; }
+        public double RequiredPoint { get; set; }
+        public double Reward { get; set; }
         public bool IsComplete { get; set; } = false;
     }
 }
