+++ PlayStudioQuestEngine/QuestEngine.Domain/Entities/Quest.cs
@@ -4,8 +4,8 @@
     {
         public string Id { get; set; }
         public string QuestTitle { get; set; }
-        public int CurrentPoint { get; set; }
-        public int TargetPoint { get; set; }
+        public double CurrentPoint { get; set; }
+        public double TargetPoint { get; set; }
         public int NumberOfMilestones { get; set; }
         public bool IsComplete { get; set; } = false;
         public bool IsActive { get; set; }
